+++ sshmitm/session.py
@@ -106,7 +106,7 @@ class Session(BaseSession):
         client_address: Union[Tuple[str, int], Tuple[str, int, int, int]],
         authenticator: Type['sshmitm.authentication.Authenticator'],
         remoteaddr: Union[Tuple[str, int], Tuple[str, int, int, int]],
-        ssh_banner: str
+        banner_name: Optional[str] = None
     ) -> None:
         """
         Initialize the class instance.
@@ -171,7 +171,7 @@ class Session(BaseSession):
 
         self.env_requests: Dict[bytes, bytes] = {}
         self.session_log_dir: Optional[str] = self.get_session_log_dir()
-        self.ssh_banner = ssh_banner
+        self.banner_name = banner_name
 
     def get_session_log_dir(self) -> Optional[str]:
         """
@@ -221,7 +221,8 @@ class Session(BaseSession):
         """
         if self._transport is None:
             self._transport = Transport(self.client_socket)
-            self.transport.local_version = f"SSH-2.0-{self.ssh_banner}"
+            if self.banner_name:
+                self.transport.local_version = f"SSH-2.0-{self.banner_name}"
             key_negotiation.handle_key_negotiation(self)
             if self.CIPHERS:
                 if not isinstance(self.CIPHERS, tuple):